requirements.txt.txt
Flask
google-cloud-storage
google-cloud-documentai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
